+++ apps/frontend/src/modules/groups/groups-api.ts
@@ -1,11 +1,15 @@
-import { APIPath } from "~/libs/enums/enums.js";
+import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
 import { type HTTP } from "~/libs/modules/http/http.js";
 import { type Storage } from "~/libs/modules/storage/storage.js";
 import { type PaginationQueryParameters } from "~/libs/types/types.js";
 
 import { GroupsApiPath } from "./libs/enums/enums.js";
-import { type GroupGetAllResponseDto } from "./libs/types/types.js";
+import {
+	type GroupCreateRequestDto,
+	type GroupCreateResponseDto,
+	type GroupGetAllResponseDto,
+} from "./libs/types/types.js";
 
 type Constructor = {
 	baseUrl: string;
@@ -17,6 +21,23 @@ class GroupApi extends BaseHTTPApi {
 	public constructor({ baseUrl, http, storage }: Constructor) {
 		super({ baseUrl, http, path: APIPath.GROUPS, storage });
 	}
+
+	public async create(
+		payload: GroupCreateRequestDto,
+	): Promise<GroupCreateResponseDto> {
+		const response = await this.load(
+			this.getFullEndpoint(GroupsApiPath.ROOT, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "POST",
+				payload: JSON.stringify(payload),
+			},
+		);
+
+		return await response.json<GroupCreateResponseDto>();
+	}
+
 	public async getAll(
 		query: PaginationQueryParameters,
 	): Promise<GroupGetAllResponseDto> {